Differentiate between screen space and world space
Also refactor to uniformly use doubles. Now we only lose precision at the boundary with NanoVG.

+++ src/main/kotlin/Extensions.kt
@@ -1,12 +1,12 @@
 inline infix fun <A, B> Pair<A, A>.map(f: (A) -> B): Pair<B, B> = f(first) to f(second)
 
-operator fun Pair<Float, Float>.plus(other: Pair<Float, Float>): Pair<Float, Float> =
+operator fun Pair<Double, Double>.plus(other: Pair<Double, Double>): Pair<Double, Double> =
     first + other.first to second + other.second
 
-operator fun Pair<Float, Float>.minus(other: Pair<Float, Float>): Pair<Float, Float> =
+operator fun Pair<Double, Double>.minus(other: Pair<Double, Double>): Pair<Double, Double> =
     first - other.first to second - other.second
 
-operator fun Pair<Float, Float>.div(other: Pair<Float, Float>): Pair<Float, Float> =
+operator fun Pair<Double, Double>.div(other: Pair<Double, Double>): Pair<Double, Double> =
     first / other.first to second / other.second
 
-operator fun Float.times(p: Pair<Float, Float>): Pair<Float, Float> = this * p.first to this * p.second
+operator fun Double.times(p: Pair<Double, Double>): Pair<Double, Double> = this * p.first to this * p.second
